# hellokratos/api/rider/v1/income_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v3.19.4
// source: rider/v1/income.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationIncomeApplyWithdrawal = "/api.rider.v1.Income/ApplyWithdrawal"
const OperationIncomeCreateIncome = "/api.rider.v1.Income/CreateIncome"
const OperationIncomeGetIncomeList = "/api.rider.v1.Income/GetIncomeList"
const OperationIncomeGetTotalIncome = "/api.rider.v1.Income/GetTotalIncome"
const OperationIncomeGetWithdrawalList = "/api.rider.v1.Income/GetWithdrawalList"
const OperationIncomeUpdateWithdrawalStatus = "/api.rider.v1.Income/UpdateWithdrawalStatus"

type IncomeHTTPServer interface {
	// ApplyWithdrawal 申请提现
	ApplyWithdrawal(context.Context, *ApplyWithdrawalRequest) (*ApplyWithdrawalReply, error)
	// CreateIncome 创建收入记录
	CreateIncome(context.Context, *CreateIncomeRequest) (*CreateIncomeReply, error)
	// GetIncomeList 获取收入列表
	GetIncomeList(context.Context, *GetIncomeListRequest) (*GetIncomeListReply, error)
	// GetTotalIncome 获取总收入
	GetTotalIncome(context.Context, *GetTotalIncomeRequest) (*GetTotalIncomeReply, error)
	// GetWithdrawalList 获取提现列表
	GetWithdrawalList(context.Context, *GetWithdrawalListRequest) (*GetWithdrawalListReply, error)
	// UpdateWithdrawalStatus 更新提现状态
	UpdateWithdrawalStatus(context.Context, *UpdateWithdrawalStatusRequest) (*UpdateWithdrawalStatusReply, error)
}

func RegisterIncomeHTTPServer(s *http.Server, srv IncomeHTTPServer) {
	r := s.Route("/")
	r.POST("/rider/v1/income/create", _Income_CreateIncome0_HTTP_Handler(srv))
	r.GET("/rider/v1/incomes", _Income_GetIncomeList0_HTTP_Handler(srv))
	r.GET("/rider/v1/income/total", _Income_GetTotalIncome0_HTTP_Handler(srv))
	r.POST("/rider/v1/withdrawal/apply", _Income_ApplyWithdrawal0_HTTP_Handler(srv))
	r.GET("/rider/v1/withdrawals", _Income_GetWithdrawalList0_HTTP_Handler(srv))
	r.POST("/rider/v1/withdrawal/update", _Income_UpdateWithdrawalStatus0_HTTP_Handler(srv))
}

func _Income_CreateIncome0_HTTP_Handler(srv IncomeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateIncomeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationIncomeCreateIncome)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateIncome(ctx, req.(*CreateIncomeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateIncomeReply)
		return ctx.Result(200, reply)
	}
}

func _Income_GetIncomeList0_HTTP_Handler(srv IncomeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetIncomeListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationIncomeGetIncomeList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetIncomeList(ctx, req.(*GetIncomeListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetIncomeListReply)
		return ctx.Result(200, reply)
	}
}

func _Income_GetTotalIncome0_HTTP_Handler(srv IncomeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetTotalIncomeRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationIncomeGetTotalIncome)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetTotalIncome(ctx, req.(*GetTotalIncomeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetTotalIncomeReply)
		return ctx.Result(200, reply)
	}
}

func _Income_ApplyWithdrawal0_HTTP_Handler(srv IncomeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ApplyWithdrawalRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationIncomeApplyWithdrawal)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ApplyWithdrawal(ctx, req.(*ApplyWithdrawalRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ApplyWithdrawalReply)
		return ctx.Result(200, reply)
	}
}

func _Income_GetWithdrawalList0_HTTP_Handler(srv IncomeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetWithdrawalListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationIncomeGetWithdrawalList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetWithdrawalList(ctx, req.(*GetWithdrawalListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetWithdrawalListReply)
		return ctx.Result(200, reply)
	}
}

func _Income_UpdateWithdrawalStatus0_HTTP_Handler(srv IncomeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateWithdrawalStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationIncomeUpdateWithdrawalStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateWithdrawalStatus(ctx, req.(*UpdateWithdrawalStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateWithdrawalStatusReply)
		return ctx.Result(200, reply)
	}
}

type IncomeHTTPClient interface {
	// ApplyWithdrawal 申请提现
	ApplyWithdrawal(ctx context.Context, req *ApplyWithdrawalRequest, opts ...http.CallOption) (rsp *ApplyWithdrawalReply, err error)
	// CreateIncome 创建收入记录
	CreateIncome(ctx context.Context, req *CreateIncomeRequest, opts ...http.CallOption) (rsp *CreateIncomeReply, err error)
	// GetIncomeList 获取收入列表
	GetIncomeList(ctx context.Context, req *GetIncomeListRequest, opts ...http.CallOption) (rsp *GetIncomeListReply, err error)
	// GetTotalIncome 获取总收入
	GetTotalIncome(ctx context.Context, req *GetTotalIncomeRequest, opts ...http.CallOption) (rsp *GetTotalIncomeReply, err error)
	// GetWithdrawalList 获取提现列表
	GetWithdrawalList(ctx context.Context, req *GetWithdrawalListRequest, opts ...http.CallOption) (rsp *GetWithdrawalListReply, err error)
	// UpdateWithdrawalStatus 更新提现状态
	UpdateWithdrawalStatus(ctx context.Context, req *UpdateWithdrawalStatusRequest, opts ...http.CallOption) (rsp *UpdateWithdrawalStatusReply, err error)
}

type IncomeHTTPClientImpl struct {
	cc *http.Client
}

func NewIncomeHTTPClient(client *http.Client) IncomeHTTPClient {
	return &IncomeHTTPClientImpl{client}
}

// ApplyWithdrawal 申请提现
func (c *IncomeHTTPClientImpl) ApplyWithdrawal(ctx context.Context, in *ApplyWithdrawalRequest, opts ...http.CallOption) (*ApplyWithdrawalReply, error) {
	var out ApplyWithdrawalReply
	pattern := "/rider/v1/withdrawal/apply"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationIncomeApplyWithdrawal))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// CreateIncome 创建收入记录
func (c *IncomeHTTPClientImpl) CreateIncome(ctx context.Context, in *CreateIncomeRequest, opts ...http.CallOption) (*CreateIncomeReply, error) {
	var out CreateIncomeReply
	pattern := "/rider/v1/income/create"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationIncomeCreateIncome))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetIncomeList 获取收入列表
func (c *IncomeHTTPClientImpl) GetIncomeList(ctx context.Context, in *GetIncomeListRequest, opts ...http.CallOption) (*GetIncomeListReply, error) {
	var out GetIncomeListReply
	pattern := "/rider/v1/incomes"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationIncomeGetIncomeList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetTotalIncome 获取总收入
func (c *IncomeHTTPClientImpl) GetTotalIncome(ctx context.Context, in *GetTotalIncomeRequest, opts ...http.CallOption) (*GetTotalIncomeReply, error) {
	var out GetTotalIncomeReply
	pattern := "/rider/v1/income/total"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationIncomeGetTotalIncome))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetWithdrawalList 获取提现列表
func (c *IncomeHTTPClientImpl) GetWithdrawalList(ctx context.Context, in *GetWithdrawalListRequest, opts ...http.CallOption) (*GetWithdrawalListReply, error) {
	var out GetWithdrawalListReply
	pattern := "/rider/v1/withdrawals"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationIncomeGetWithdrawalList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateWithdrawalStatus 更新提现状态
func (c *IncomeHTTPClientImpl) UpdateWithdrawalStatus(ctx context.Context, in *UpdateWithdrawalStatusRequest, opts ...http.CallOption) (*UpdateWithdrawalStatusReply, error) {
	var out UpdateWithdrawalStatusReply
	pattern := "/rider/v1/withdrawal/update"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationIncomeUpdateWithdrawalStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}