# hellokratos/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /rider/v1/income/create:
        post:
            tags:
                - Income
            description: 创建收入记录
            operationId: Income_CreateIncome
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.CreateIncomeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.CreateIncomeReply'
    /rider/v1/income/total:
        get:
            tags:
                - Income
            description: 获取总收入
            operationId: Income_GetTotalIncome
            parameters:
                - name: riderId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.GetTotalIncomeReply'
    /rider/v1/incomes:
        get:
            tags:
                - Income
            description: 获取收入列表
            operationId: Income_GetIncomeList
            parameters:
                - name: riderId
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.GetIncomeListReply'
...
    /rider/v1/withdrawal/apply:
        post:
            tags:
                - Income
            description: 申请提现
            operationId: Income_ApplyWithdrawal
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.ApplyWithdrawalRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.ApplyWithdrawalReply'
    /rider/v1/withdrawal/update:
        post:
            tags:
                - Income
            description: 更新提现状态
            operationId: Income_UpdateWithdrawalStatus
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.UpdateWithdrawalStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.UpdateWithdrawalStatusReply'
    /rider/v1/withdrawals:
        get:
            tags:
                - Income
            description: 获取提现列表
            operationId: Income_GetWithdrawalList
            parameters:
                - name: riderId
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.GetWithdrawalListReply'
components:
    schemas:
...
        api.rider.v1.ApplyWithdrawalReply:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
                withdrawalId:
                    type: string
            description: 申请提现响应
        api.rider.v1.ApplyWithdrawalRequest:
            type: object
            properties:
                riderId:
                    type: string
                amount:
                    type: number
                    format: float
                platform:
                    type: string
                account:
                    type: string
            description: 申请提现请求
        api.rider.v1.CreateIncomeReply:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
                incomeId:
                    type: string
                amount:
                    type: number
                    format: float
                distance:
                    type: number
                    format: float
            description: 创建收入记录响应
        api.rider.v1.CreateIncomeRequest:
            type: object
            properties:
                orderId:
                    type: string
                incomeType:
                    type: integer
                    format: int32
                incomeStatus:
                    type: integer
                    format: int32
            description: 创建收入记录请求
...
        api.rider.v1.GetIncomeListReply:
            type: object
            properties:
                incomes:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.rider.v1.IncomeInfo'
                total:
                    type: integer
                    format: int32
            description: 获取收入列表响应
...
        api.rider.v1.GetTotalIncomeReply:
            type: object
            properties:
                total:
                    type: number
                    format: float
            description: 获取总收入响应
...
        api.rider.v1.GetWithdrawalListReply:
            type: object
            properties:
                withdrawals:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.rider.v1.WithdrawalInfo'
                total:
                    type: integer
                    format: int32
            description: 获取提现列表响应
        api.rider.v1.IncomeInfo:
            type: object
            properties:
                id:
                    type: string
                riderId:
                    type: string
                orderId:
                    type: string
                amount:
                    type: number
                    format: float
                incomeType:
                    type: integer
                    format: int32
                incomeStatus:
                    type: integer
                    format: int32
                createdAt:
                    type: string
            description: 收入信息
...
        api.rider.v1.UpdateWithdrawalStatusReply:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
            description: 更新提现状态响应
        api.rider.v1.UpdateWithdrawalStatusRequest:
            type: object
            properties:
                withdrawalId:
                    type: string
                status:
                    type: integer
                    format: int32
                transactionId:
                    type: string
            description: 更新提现状态请求
...
        api.rider.v1.WithdrawalInfo:
            type: object
            properties:
                id:
                    type: string
                riderId:
                    type: string
                amount:
                    type: number
                    format: float
                platform:
                    type: string
                account:
                    type: string
                status:
                    type: integer
                    format: int32
                transactionId:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
            description: 提现信息
...